apps/ml-service/pyproject.toml
[project]
name = "lumo-ml-service"
version = "0.1.0"
description = "Python-backed Lumo Intelligence Layer exposed as a privileged Lumo system agent."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "e2b-code-interpreter==2.5.0",
  "fastapi>=0.115.0",
  "ortools>=9.15.6755",
  "modal>=1.0.0",
  "presidio-analyzer>=2.2.359",
  "presidio-anonymizer>=2.2.359",
  "pydantic>=2.9.0",
  "PyJWT>=2.9.0",
  "unstructured[pdf]>=0.16.0",
  "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.27.0",
  "pytest>=8.3.0",
  "ruff>=0.6.0",
  "mypy>=1.11.0",
  "pydantic-to-typescript>=2.0.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["lumo_ml*", "app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
follow_imports = "silent"
warn_unused_ignores = false
# Pydantic v2 + Literal-typed fields produce many false positives at the
# call site (`status="ok"` flagged because `str` doesn't narrow to the
# Literal union). The codegen + drift check defends the Pydantic source of
# truth; we keep mypy on for real reference/import errors but suppress
# narrowing noise. Modal SDK + transformers-based modules are also dynamic
# enough that `arg-type`/`return-value` checks here aren't useful.
disable_error_code = ["arg-type", "return-value", "misc"]
